feat: Documented + cleared the function vector code, now registration is working in cli and the fn_compute_vect_ allow deletion and function tracking by ids.

+++ Holovibes/includes/tools/aliases.hh
@@ -25,7 +25,7 @@ using ConditionType = std::function<bool()>;
 using FnType = std::function<void()>;
 
 /*! \brief A procedure vector. */
-using FnVector = std::vector<std::pair<int, FnType>>;
+using FnVector = std::vector<std::pair<ushort, FnType>>;
 
 /*! \brief A procedure deque. */
 using FnDeque = std::deque<FnType>;
