+++ Holovibes/sources/compute/registration.cc
@@ -16,12 +16,6 @@ void Registration::insert_registration()
         fn_compute_vect_.conditional_push_back(
             [=]()
             {
-                // cudaXMemcpyAsync(buffers_.gpu_postprocess_frame,
-                //                  gpu_reference_image_,
-                //                  fd_.width * fd_.height * sizeof(float),
-                //                  cudaMemcpyDeviceToDevice,
-                //                  stream_);
-                // return;
                 // Preprocessing the current image before the cross-correlation with the reference image.
                 image_preprocess(gpu_current_image_, buffers_.gpu_postprocess_frame, &current_image_mean_);
 
@@ -75,7 +69,7 @@ void Registration::set_gpu_reference_image(float* new_gpu_reference_image_)
 {
     if (setting<settings::RegistrationEnabled>())
     {
-        int func_id = fn_compute_vect_.conditional_push_back(
+        ushort func_id = fn_compute_vect_.conditional_push_back(
             [=]
             {
                 cudaXMemcpyAsync(gpu_reference_image_,
@@ -86,6 +80,10 @@ void Registration::set_gpu_reference_image(float* new_gpu_reference_image_)
 
                 image_preprocess(gpu_reference_image_, gpu_reference_image_, &reference_image_mean_);
             });
+
+        // After the `gpu_accumulation_xy_queue` buffer is full, we have a reference image on a fully accumulated image.
+        // Then we can remove this function from the `fn_compute_vect_`, since we consider having a good enough
+        // reference.
         fn_compute_vect_.conditionnal_remove(func_id,
                                              [this]
                                              {
